requirements.txt
google-genai
python-jobspy
pandas
python-dotenv